refactor: Adds error checks while fetching
