+++ ui/src/app/enums/internal-route.ts
